Interlok 3571 (#71)
* INTERLOK-3571 Now a multi-module project.

* INTERLOK-3571 Java 11'd and split the project into a multi module project.

* INTERLOK-3571 Removed the base set of src files now they have been split.
