+++ interlok-rest-base/src/main/resources/META-INF/com/adaptris/core/management/components/rest
