+++ interlok-rest-base/src/test/java/com/adaptris/rest/MockWorkflowConsumer.java
