+++ interlok-rest-base/src/test/java/com/adaptris/rest/util/Dummy.java
