+++ interlok-rest-cluster/src/main/resources/META-INF/com/adaptris/core/management/components/cluster-rest
