+++ interlok-rest-health-check/src/main/java/com/adaptris/rest/healthcheck/State.java
