+++ interlok-rest-health-check/src/main/java/com/adaptris/rest/healthcheck/WorkflowState.java
@@ -1,0 +1,9 @@
+package com.adaptris.rest.healthcheck;
+
+import com.thoughtworks.xstream.annotations.XStreamAlias;
+import lombok.NoArgsConstructor;
+
+@XStreamAlias("workflow-state")
+@NoArgsConstructor
+public class WorkflowState extends State {
+}
