+++ interlok-rest-health-check/src/main/resources/META-INF/com/adaptris/core/management/components/health-check
