+++ interlok-rest-health-check/src/test/java/com/adaptris/rest/healthcheck/ChannelStateTest.java
