+++ interlok-rest-prometheus/src/main/resources/META-INF/com/adaptris/core/management/components/prometheus-rest
