+++ lombok.config
@@ -2,5 +2,3 @@
 config.stopBubbling = true
 lombok.addLombokGeneratedAnnotation = true
 lombok.extern.findbugs.addSuppressFBWarnings = true
-lombok.getter.noIsPrefix = true
-lombok.copyableAnnotations += com.adaptris.annotation.Removal
